pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "repo_runner"
version = "0.1.0"
description = "AI-enabled modular CLI agent for analyzing repos, auto-installing dependencies, and running applications"
authors = [
    {name = "AI Assistant", email = "assistant@example.com"}
]
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]
dependencies = [
    "click>=8.0.0",
    "transformers>=4.30.0",
    "torch>=2.0.0",
    "accelerate>=0.20.0",
    "requests>=2.28.0",
    "psutil>=5.9.0",
    "colorama>=0.4.6",
    "rich>=13.0.0",
    "pyyaml>=6.0",
    "jinja2>=3.0.0",
]

[project.scripts]
repo_runner = "repo_runner.cli:cli"

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["repo_runner*"]

[tool.black]
line-length = 88
target-version = ['py38']

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"] 